Document the current status of LibreCMC and LEDE.
From browsing the GnuBee website, wiki pages, Github issues and Google Group, my
understanding is that LibreCMC and LEDE are not actively supported. I think it
is very important to be clear about the status of these two options and also to
provide clear recommendations of what is being supported.